requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
pypdf2
faiss-cpu
sentence-transformers
python-dotenv